+++ CHANGELOG.md
@@ -1,5 +1,7 @@
+
 ## 0.1.0
 
 Features:
 
   - Created the cookbook `php52`
+  - Added the PHP5.2.17 installation with httpd